backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
torch
transformers
accelerate
huggingface_hub